Fix X020 large-corpus parity
