+++ crates/shuck-linter/src/rules/portability/brace_fd_redirection.rs
@@ -1,5 +1,3 @@
-use shuck_ast::Span;
-
 use crate::{Checker, Rule, ShellDialect, Violation};
 
 pub struct BraceFdRedirection;
@@ -24,38 +22,10 @@ pub fn brace_fd_redirection(checker: &mut Checker) {
         .commands()
         .iter()
         .flat_map(|fact| fact.redirect_facts().iter())
-        .filter_map(|redirect| brace_fd_span(redirect.redirect(), checker.source()))
+        .filter_map(|redirect| redirect.brace_fd_redirection_span())
         .collect::<Vec<_>>();
 
     checker.report_all_dedup(spans, || BraceFdRedirection);
-}
-
-fn brace_fd_span(redirect: &shuck_ast::Redirect, source: &str) -> Option<Span> {
-    let fd_var_span = redirect.fd_var_span?;
-    let brace_close = fd_var_span.end.advanced_by("}");
-    let gap = &source[brace_close.offset..redirect.span.start.offset];
-    brace_fd_gap_allows_attachment(gap).then_some(fd_var_span)
-}
-
-fn brace_fd_gap_allows_attachment(gap: &str) -> bool {
-    if gap.is_empty() {
-        return true;
-    }
-
-    let mut remaining = gap;
-    while !remaining.is_empty() {
-        if let Some(stripped) = remaining.strip_prefix("\\\r\n") {
-            remaining = stripped;
-            continue;
-        }
-        if let Some(stripped) = remaining.strip_prefix("\\\n") {
-            remaining = stripped;
-            continue;
-        }
-        return false;
-    }
-
-    true
 }
 
 #[cfg(test)]
@@ -79,9 +49,9 @@ EOF
         let diagnostics = test_snippet(source, &LinterSettings::for_rule(Rule::BraceFdRedirection));
 
         assert_eq!(diagnostics.len(), 3);
-        assert_eq!(diagnostics[0].span.slice(source), "fd");
-        assert_eq!(diagnostics[1].span.slice(source), "docfd");
-        assert_eq!(diagnostics[2].span.slice(source), "contfd");
+        assert_eq!(diagnostics[0].span.slice(source), "{fd}>/dev/null");
+        assert_eq!(diagnostics[1].span.slice(source), "{docfd}<<EOF");
+        assert_eq!(diagnostics[2].span.slice(source), "{contfd}<<EOF");
     }
 
     #[test]
